// src/ts/enums/store/user.enum.ts
export enum EUserActions {
    setTotal = "setTotal",
    setUsers = "setUsers",
    addUser = "addUser",
    updateUser = "updateUser",
    deleteUser = "deleteUser",
}

export enum EUserGetters {
    getTotalUsers = "getTotalUsers",
    getUsersByPage = "getUsersByPage",
    getUserById = "getUserById",
    getUsersBySearch = "getUsersBySearch",
}